+++ dompdf/include/image_frame_decorator.cls.php
@@ -35,9 +35,14 @@
  * @author Benj Carson <benjcarson@digitaljunkies.ca>
  * @package dompdf
  * @version 0.5.1
+ *
+ * Changes
+ * @author Helmut Tischer <htischer@weihenstephan.org>
+ * @version 0.5.1.htischer.20090507
+ * - add optional debug output
  */
 
-/* $Id: image_frame_decorator.cls.php,v 1.12 2006-08-02 18:44:25 benjcarson Exp $ */
+/* $Id: image_frame_decorator.cls.php,v 1.11 2006-07-07 21:31:03 benjcarson Exp $ */
 
 /**
  * Decorates frames for image layout and rendering
@@ -81,7 +86,10 @@ class Image_Frame_Decorator extends Frame_Decorator {
     
     parent::__construct($frame, $dompdf);
     $url = $frame->get_node()->getAttribute("src");
-      
+
+    //debugpng
+    if (DEBUGPNG) print '[__construct '.$url.']';
+
     list($this->_image_url, $this->_image_ext) = Image_Cache::resolve_url($url,
                                                                           $dompdf->get_protocol(),
                                                                           $dompdf->get_host(),
@@ -114,7 +122,10 @@ class Image_Frame_Decorator extends Frame_Decorator {
   static function clear_image_cache() {
     if ( count(self::$_cache) ) {
       foreach (self::$_cache as $file)
-        unlink($file);
+        //debugpng
+        if (DEBUGPNG) print '[clear_image_cache unlink '.$file.']';
+        if (!DEBUGKEEPTEMP)
+          unlink($file);
     }
   }
 }
